Move metrics tracker into shared reporting utilities
- Relocate createMetricsTracker from src/plugin/src/reporting/metrics-tracker.js to src/shared/reporting/metrics-tracker.js so instrumentation lives in the shared utilities layer shared by plugin features.

- Expose the helper through shared/reporting barrels, update plugin imports, and leave a compatibility re-export under src/plugin/src/reporting/metrics-tracker.js for existing entry points.
